api_yamdb/requirements.txt
Django==3.2.*
PyJWT==2.1.0
asgiref==3.3.2
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1